Migrates darwinia-network/bridge-primitives
