+++ Cargo.toml
@@ -1,51 +1,60 @@
+[workspace]
+members = [
+  "primitives",
+]
+
 [package]
-name = "darwinia-bridge-primitives"
-version = "0.0.12"
+name = "darwinia-bridger"
+version = "0.0.3"
 authors = ["clearloop <udtrokia@gmail.com>"]
-description = "Darwinia bridge primitives"
-repository = "https://github.com/darwinia-network/bridge-primitives"
-license = "GPL-3.0"
-documentation = "https://docs.rs/darwinia-bridge-primitives"
-homepage = "https://github.com/darwinia-network/bridge-primitives"
+description = "Darwinia bridger"
+repository = "https://github.com/darwinia-network/bridger"
+license = "MIT"
+documentation = "https://docs.rs/bridger"
+homepage = "https://github.com/darwinia-network/bridger"
 include = ["src/**/*", "Cargo.toml", "./README.md"]
-keywords = ["blockchain"]
+keywords = ["substrate", "darwinia"]
 readme = './README.md'
 edition = "2018"
 
-[package.metadata.docs.rs]
-all-features = true
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "bridger"
+path = "src/bin/bridger.rs"
 
 [dependencies]
-uint = "0.8.5"
-rlp = "0.4.5"
+async-trait = "0.1.40"
+dirs = "3.0.1"
+env_logger = "0.7.1"
+etc = "0.1.11"
+futures = "0.3.5"
+log = "0.4.11"
+sp-keyring = "2.0.0"
+structopt = "0.3.18"
+substrate-subxt = "0.13.0"
+toml = "0.5.6"
+web3 = "0.13.0"
+serde_json = "1.0"
+hex = "0.4.2"
 
-## Optional featrues
-async-trait = { version = "0.1.40", optional = true }
-frame-support = { version = "2.0.0", optional = true }
-pallet-indices = { version = "2.0.0", optional = true }
-pallet-im-online = { version = "2.0.0", optional = true }
-reqwest = { version = "0.10", features = ["json"], optional = true }
-serde_json = { version = "1.0", optional = true }
-substrate-subxt = { version = "0.13.0", optional = true }
-substrate-subxt-proc-macro = { version = "0.13.0", optional = true }
+[dependencies.reqwest]
+version = "0.10"
+features = ["json"]
+
+[dependencies.tokio]
+package = "tokio"
+version = "0.2.22"
+features = ["full"]
+
+[dependencies.primitives]
+package = "darwinia-bridge-primitives"
+version = "0.0.12"
+path = "./primitives"
+features = ["runtime"]
 
 [dependencies.serde]
 package = "serde"
 version = "1"
 features = ["derive"]
-
-[dependencies.codec]
-package = "parity-scale-codec"
-version = "1"
-features = ["derive"]
-
-[features]
-rpc = ["async-trait", "reqwest", "serde_json"]
-runtime = [
-  "substrate-subxt",
-  "substrate-subxt-proc-macro",
-  "frame-support",
-  "pallet-indices",
-  "pallet-im-online",
-]
-default = []
