--- crates/bridger/README.md
@@ -1,38 +1,0 @@
-## Bridger
-
-[![bridger](https://github.com/darwinia-network/bridger/workflows/bridger/badge.svg)](https://github.com/darwinia-network/bridger)
-[![crate](https://img.shields.io/crates/v/darwinia-bridger.svg)](https://crates.io/crates/darwinia-bridger)
-[![doc](https://img.shields.io/badge/current-docs-brightgreen.svg)](https://docs.rs/darwinia-bridger/)
-[![downloads](https://img.shields.io/crates/d/darwinia-bridger.svg)](https://crates.io/crates/darwinia-bridger)
-[![LICENSE](https://img.shields.io/crates/l/darwinia-bridger.svg)](https://choosealicense.com/licenses/gpl/)
-
-The Darwinia Bridger
-
-## Installation
-
-```
-$ cargo install darwinia-bridger
-```
-
-## Usage
-
-```
-$ bridger
-darwinia-bridger 0.0.3
-
-USAGE:
-    bridger <SUBCOMMAND>
-
-FLAGS:
-    -h, --help       Prints help information
-    -V, --version    Prints version information
-
-SUBCOMMANDS:
-    confirm    Set Confirmed block with sudo privilege
-    help       Prints this message or the help of the given subcommand(s)
-    run        Run the bridger
-```
-
-## LICENSE
-
-GPL-3.0
