--- crates/bridger/src/lib.rs
@@ -1,16 +1,0 @@
-//! Darwinia bridger
-#![warn(missing_docs)]
-
-#[macro_use]
-extern crate log;
-
-mod config;
-mod listener;
-
-pub mod api;
-pub mod cmd;
-pub mod pool;
-pub mod result;
-pub mod service;
-
-pub use self::{config::Config, listener::Listener, pool::Pool, service::Service};
