+++ primitives/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "darwinia-bridge-primitives"
+version = "0.0.12"
+authors = ["clearloop <udtrokia@gmail.com>"]
+description = "Darwinia bridge primitives"
+repository = "https://github.com/darwinia-network/bridge-primitives"
+license = "GPL-3.0"
+documentation = "https://docs.rs/darwinia-bridge-primitives"
+homepage = "https://github.com/darwinia-network/bridge-primitives"
+include = ["src/**/*", "Cargo.toml", "./README.md"]
+keywords = ["blockchain"]
+readme = './README.md'
+edition = "2018"
+
+[package.metadata.docs.rs]
+all-features = true
+
+[dependencies]
+uint = "0.8.5"
+rlp = "0.4.5"
+
+## Optional featrues
+async-trait = { version = "0.1.40", optional = true }
+frame-support = { version = "2.0.0", optional = true }
+pallet-indices = { version = "2.0.0", optional = true }
+pallet-im-online = { version = "2.0.0", optional = true }
+reqwest = { version = "0.10", features = ["json"], optional = true }
+serde_json = { version = "1.0", optional = true }
+substrate-subxt = { version = "0.13.0", optional = true }
+substrate-subxt-proc-macro = { version = "0.13.0", optional = true }
+
+[dependencies.serde]
+package = "serde"
+version = "1"
+features = ["derive"]
+
+[dependencies.codec]
+package = "parity-scale-codec"
+version = "1"
+features = ["derive"]
+
+[features]
+rpc = ["async-trait", "reqwest", "serde_json"]
+runtime = [
+  "substrate-subxt",
+  "substrate-subxt-proc-macro",
+  "frame-support",
+  "pallet-indices",
+  "pallet-im-online",
+]
+default = []
