+++ primitives/src/array.rs
