+++ src/cmd/affirm.rs
