+++ src/result.rs
@@ -1,19 +1,20 @@
-//! Bridge Result
+//! Bridger Result
+use etc::Error as Etc;
+use reqwest::Error as Reqwest;
+use serde_json::Error as SerdeJson;
 use std::{
     error::Error as ErrorTrait,
     fmt::{Display, Formatter, Result as FmtResult},
     io::Error as Io,
     result::Result as StdResult,
 };
-
-#[cfg(feature = "rpc")]
-use reqwest::Error as Reqwest;
-#[cfg(feature = "rpc")]
-use serde_json::Error as SerdeJson;
+use substrate_subxt::Error as Subxt;
+use toml::{de::Error as DeToml, ser::Error as SerToml};
+use web3::Error as Web3;
 
 /// The custom bridger error
-pub struct Bridge(String);
-impl Display for Bridge {
+pub struct Bridger(String);
+impl Display for Bridger {
     fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
         f.write_str(&self.0)
     }
@@ -21,43 +22,35 @@ impl Display for Bridge {
 
 /// Error generator
 macro_rules! error {
-    ($($(#[$attr:meta])* $e:ident),*) => {
-        /// Bridge Error
+    ($($e:ident),*) => {
+        /// Bridger Error
         #[derive(Debug)]
         #[allow(missing_docs)]
         pub enum Error {
-            $($(#[$attr])* $e(String),)+
+            $($e(String),)+
         }
 
         impl Display for Error {
             fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
                 match self {
-                    $($(#[$attr])* Error::$e(e) => e.fmt(f),)+
+                    $(Error::$e(e) => e.fmt(f),)+
                 }
             }
         }
 
         impl ErrorTrait for Error {}
+
         $(
-            $(#[$attr])*
             impl From<$e> for Error {
                 fn from(e: $e) -> Error {
                     Error::$e(format!("{}", e))
                 }
             }
-        )*
-
+        )+
     };
 }
 
-error! {
-    Io,
-    Bridge,
-    #[cfg(feature = "rpc")]
-    Reqwest,
-    #[cfg(feature = "rpc")]
-    SerdeJson
-}
+error! {Io, Bridger, DeToml, SerToml, Etc, Web3, Reqwest, SerdeJson, Subxt}
 
-/// Sup Result
+/// Bridger Result
 pub type Result<T> = StdResult<T, Error>;
